feat: title-based JOIN, text table sync, INFINITAS song filtering
Major changes to web sync architecture:
- Replace song_id-based JOIN with title+difficulty JOIN between charts and lamps
- Read game_id and V3 title from text table (stride 0x630) for sync
- Add infinitas_title column to charts, title column to lamps
- Filter non-INFINITAS songs using iidxapi music list
- Persist interactive normalize resolutions to resolutions.json
- Merge build-tracker into normalize, add justfile recipes
- Read embedded EX scores/lamps from entry table for export
- Trim trailing whitespace from Shift-JIS decoded titles
- Relax songId validation (title is the primary key now)

+++ crates/infst/src/play/play_data.rs
@@ -36,7 +36,7 @@ impl PlayData {
         if total_notes == 0 {
             return Grade::F;
         }
-        let max_ex = total_notes * 2;
+        let max_ex = total_notes as u64 * 2;
         let ratio = ex_score as f64 / max_ex as f64;
         Grade::from_score_ratio(ratio)
     }
